feat: add a polling timeout
The property timeout-minutes in a GitHub step is not supported in composed actions. So the workaround to stop the waiting for finishing the sonar scan is not working in such cases.
To fix this issue, a polling timeout variable was introduced, which stops the waiting, if the timout is reached.